[fix](doc/bmlib):update readme and bmlib pdf [description ]:NA [root   cause]:NA [JIRA      ID]:NA [chip project]:NA [side effects]:NA [Affected *.ko/*.so/*.a] :NA

+++ bmlib/src/bmcpu_app.cpp
@@ -453,7 +453,7 @@ void* bmcpu_thread(void* arg) {
       //A timer to record the API usage time
       auto start = std::chrono::high_resolution_clock::now();
 
-      switch (bm_api->api_id) {
+      switch (static_cast<unsigned int>(bm_api->api_id)) {
       case API_ID_SGTPUV8_LOAD_LIB:
           bmlib_log(BMLIB_bmcpu_LOG_TAG, BMLIB_LOG_INFO, "===========load lib!===========\n");
           api=(bm_api_cpu_load_library_internal_t *)bm_api->api_data;
